Change packaging to separate workflow interfaces from implementations. Remove paramaterization from Workflow interfaces.
