+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/context/BrokerRequestContext.java
@@ -14,19 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.appbroker.workflow.action.createserviceinstance.appdeploy;
+package org.springframework.cloud.appbroker.context;
 
-import java.util.Map;
-
-public class BackingAppState {
-
-	private Map<String, String> state;
-
-	public BackingAppState(Map<String, String> state) {
-		this.state = state;
-	}
-
-	public Map<String, String> getState() {
-		return state;
-	}
+public interface BrokerRequestContext {
 }
