+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/instance/AppDeploymentCreateServiceInstanceWorkflow.java
@@ -14,27 +14,26 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.appbroker.workflow;
+package org.springframework.cloud.appbroker.instance;
 
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cloud.appbroker.workflow.action.createserviceinstance.CreateServiceRequestContext;
-import org.springframework.cloud.appbroker.workflow.action.createserviceinstance.DefaultCreateServiceBrokerResponseBuilder;
-import org.springframework.cloud.appbroker.workflow.action.createserviceinstance.appdeploy.BackingAppDeployer;
-import org.springframework.cloud.appbroker.workflow.action.createserviceinstance.appdeploy.BackingAppDeploymentPlan;
-import org.springframework.cloud.appbroker.workflow.action.createserviceinstance.appdeploy.BackingAppParameters;
+import org.springframework.cloud.appbroker.instance.create.CreateServiceRequestContext;
+import org.springframework.cloud.appbroker.instance.create.DefaultCreateServiceBrokerResponseBuilder;
+import org.springframework.cloud.appbroker.instance.create.appdeploy.BackingAppDeploymentPlan;
+import org.springframework.cloud.appbroker.workflow.CreateServiceInstanceWorkflow;
 import org.springframework.cloud.servicebroker.model.instance.CreateServiceInstanceRequest;
 import org.springframework.cloud.servicebroker.model.instance.CreateServiceInstanceResponse;
 
-public class CreateServiceInstanceDefaultWorkflow implements CreateServiceInstanceWorkflow<CreateServiceInstanceRequest, CreateServiceInstanceResponse> {
+public class AppDeploymentCreateServiceInstanceWorkflow implements CreateServiceInstanceWorkflow {
 
 	private Set<BackingAppDeploymentPlan> deploymentPlans;
 	private DefaultCreateServiceBrokerResponseBuilder createServiceBrokerResponseBuilder;
 
 	@Autowired
-	public CreateServiceInstanceDefaultWorkflow(Set<BackingAppDeploymentPlan> deploymentPlans,
-												DefaultCreateServiceBrokerResponseBuilder createServiceBrokerResponseBuilder) {
+	public AppDeploymentCreateServiceInstanceWorkflow(Set<BackingAppDeploymentPlan> deploymentPlans,
+													  DefaultCreateServiceBrokerResponseBuilder createServiceBrokerResponseBuilder) {
 		this.deploymentPlans = deploymentPlans;
 		this.createServiceBrokerResponseBuilder = createServiceBrokerResponseBuilder;
 	}
