+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/instance/AppDeploymentGetLastOperationWorkflow.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.appbroker.workflow;
+package org.springframework.cloud.appbroker.instance;
 
+import org.springframework.cloud.appbroker.workflow.GetLastOperationWorkflow;
 import org.springframework.cloud.servicebroker.model.instance.GetLastServiceOperationRequest;
 import org.springframework.cloud.servicebroker.model.instance.GetLastServiceOperationResponse;
 
-public class GetLastOperationDefaultWorkflow implements GetLastOperationWorkflow<GetLastServiceOperationRequest, GetLastServiceOperationResponse> {
+public class AppDeploymentGetLastOperationWorkflow implements GetLastOperationWorkflow {
 
 	@Override
 	public GetLastServiceOperationResponse perform(GetLastServiceOperationRequest requestData) {
