+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/instance/create/appdeploy/CloudFoundryDefaultBackingAppDeployer.java
@@ -14,26 +14,18 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.appbroker.workflow.action.createserviceinstance.appdeploy;
+package org.springframework.cloud.appbroker.instance.create.appdeploy;
 
-import java.util.Map;
+import java.util.Collections;
 
-public class BackingAppParameters {
+import org.springframework.cloud.appbroker.instance.create.CreateServiceRequestContext;
 
-	private final String name;
+public class CloudFoundryDefaultBackingAppDeployer implements BackingAppDeployer {
 
-	private final Map<String, String> properties;
-
-	public BackingAppParameters(String name, Map<String, String> properties) {
-		this.name = name;
-		this.properties = properties;
+	@Override
+	public void accept(BackingAppParameters backingAppParameters, CreateServiceRequestContext createServiceRequestContext) {
+		//Deploy app and store any data generated by the platform that might need to be persisted, such as app guid or generated route
+		createServiceRequestContext.setBackingAppState(new BackingAppState(Collections.emptyMap()));
 	}
 
-	public String getName() {
-		return name;
-	}
-
-	public Map<String, String> getProperties() {
-		return properties;
-	}
 }
