+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/workflow/CreateServiceInstanceWorkflow.java
@@ -19,6 +19,6 @@ package org.springframework.cloud.appbroker.workflow;
 import org.springframework.cloud.servicebroker.model.instance.CreateServiceInstanceRequest;
 import org.springframework.cloud.servicebroker.model.instance.CreateServiceInstanceResponse;
 
-public interface CreateServiceInstanceWorkflow<REQ extends CreateServiceInstanceRequest, RES extends CreateServiceInstanceResponse>
-	extends AppBrokerWorkflow<REQ, RES> {
+public interface CreateServiceInstanceWorkflow
+	extends ServiceBrokerWorkflow<CreateServiceInstanceRequest, CreateServiceInstanceResponse> {
 }
