+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/workflow/DeleteServiceInstanceWorkflow.java
@@ -19,6 +19,6 @@ package org.springframework.cloud.appbroker.workflow;
 import org.springframework.cloud.servicebroker.model.instance.DeleteServiceInstanceRequest;
 import org.springframework.cloud.servicebroker.model.instance.DeleteServiceInstanceResponse;
 
-public interface DeleteServiceInstanceWorkflow<REQ extends DeleteServiceInstanceRequest, RES extends DeleteServiceInstanceResponse>
-	extends AppBrokerWorkflow<REQ, RES> {
+public interface DeleteServiceInstanceWorkflow
+	extends ServiceBrokerWorkflow<DeleteServiceInstanceRequest, DeleteServiceInstanceResponse> {
 }
