+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/workflow/GetLastOperationWorkflow.java
@@ -19,6 +19,6 @@ package org.springframework.cloud.appbroker.workflow;
 import org.springframework.cloud.servicebroker.model.instance.GetLastServiceOperationRequest;
 import org.springframework.cloud.servicebroker.model.instance.GetLastServiceOperationResponse;
 
-public interface GetLastOperationWorkflow<REQ extends GetLastServiceOperationRequest, RES extends GetLastServiceOperationResponse>
-	extends AppBrokerWorkflow<REQ, RES> {
+public interface GetLastOperationWorkflow
+	extends ServiceBrokerWorkflow<GetLastServiceOperationRequest, GetLastServiceOperationResponse> {
 }
