+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/workflow/GetServiceInstanceWorkflow.java
@@ -19,6 +19,6 @@ package org.springframework.cloud.appbroker.workflow;
 import org.springframework.cloud.servicebroker.model.instance.GetServiceInstanceRequest;
 import org.springframework.cloud.servicebroker.model.instance.GetServiceInstanceResponse;
 
-public interface GetServiceInstanceWorkflow<REQ extends GetServiceInstanceRequest, RES extends GetServiceInstanceResponse>
-	extends AppBrokerWorkflow<REQ, RES> {
+public interface GetServiceInstanceWorkflow
+	extends ServiceBrokerWorkflow<GetServiceInstanceRequest, GetServiceInstanceResponse> {
 }
